pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maskopy-local"
version = "0.1.0"
description = "Local implementation of core masking algorithms inspired by the official Maskopy framework."
requires-python = ">=3.8"
dependencies = ["mysql-connector-python"]

[tool.setuptools]
packages = ["maskopy"]
